[swap_device] change lseek return type from int to off_t
 - Return value of lseek is offset of requested position. The 4byte int type cannot represent
   more than 3GB so any request more than 3GB offset cause overflow.

+++ nntrainer/tensor/swap_device.h
@@ -81,7 +81,7 @@ public:
    * @return The pointer of the swap space
    *
    */
-  void *getBuffer(int offset, size_t size);
+  void *getBuffer(off_t offset, size_t size);
 
   /**
    * @brief Deallocate and put data
@@ -121,7 +121,7 @@ private:
   std::map<void *, std::pair<void *, size_t>>
     mapped; /**< <pointer, <orig_pointer, size>> */
 #else
-  std::map<void *, std::pair<int, ssize_t>>
+  std::map<void *, std::pair<off_t, ssize_t>>
     allocated; /**< <pointer, <offset, size>> */
 #endif
 };
